requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.3
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.27.0
